+++ src/Components/BasicButton/BasicButton.module.scss
@@ -1,0 +1,28 @@
+.basic_button {
+  margin: 0;
+  padding: 0;
+  width: 176px;
+  height: 44px;
+  font-family: Verdana;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 14px;
+  line-height: 17px;
+  text-align: center;
+  letter-spacing: 0.04em;
+  color: #ffffff;
+  background: #fc842d;
+  box-shadow: 0px 4px 10px rgba(252, 132, 45, 0.5);
+  border: 2px solid #fc842d;
+  border-radius: 30px;
+  cursor: pointer;
+}
+
+.basic_button:active,
+.basic_button:focus {
+  outline: none;
+  background: #ffffff;
+  color: #fc842d;
+  box-sizing: border-box;
+  box-shadow: none;
+}
