+++ src/Components/CalculatorCalorieForm/CalculatorCalorieForm.module.scss
@@ -3,9 +3,6 @@
   margin: 0 auto;
 }
 
-.image {
-  display: none;
-}
 .title {
   margin-top: 30px;
   margin-bottom: 30px;
@@ -59,10 +56,7 @@
   .container {
     width: 100%;
   }
-  .section {
-    width: auto;
-    height: 850px;
-  }
+
   .title {
     font-size: 30px;
     line-height: 37.5px;
@@ -89,35 +83,6 @@
   .left_input {
     display: block;
     margin-right: 30px;
-  }
-  .container {
-    position: relative;
-  }
-  .image {
-    display: block;
-  }
-  .image_box {
-    position: absolute;
-    width: 602px;
-    height: 816px;
-  }
-  .overlay_strawberry {
-    position: absolute;
-    z-index: 3;
-    // right: ;
-    // bottom: ;
-  }
-  .overlay_bannana {
-    position: absolute;
-    z-index: 3;
-  }
-  .overlay_white {
-    position: absolute;
-    z-index: 1;
-  }
-  .overlay_leaves {
-    position: absolute;
-    z-index: 5;
   }
 }
